+++ Jellyfin.Plugin.JellyRock/PluginServiceRegistrator.cs
@@ -16,5 +16,6 @@ public sealed class PluginServiceRegistrator : IPluginServiceRegistrator
     public void RegisterServices(IServiceCollection serviceCollection, IServerApplicationHost applicationHost)
     {
         serviceCollection.AddHostedService<JellyRockSessionService>();
+        serviceCollection.AddHostedService<PlaybackReaperService>();
     }
 }